new set

+++ src/main/java/studio/baivo/testenv_sakila/Repository/FilmRepository.java
@@ -3,6 +3,7 @@ package studio.baivo.testenv_sakila.Repository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
+import studio.baivo.testenv_sakila.model.Film;
 
 import java.util.Set;
 
@@ -10,7 +11,7 @@ import java.util.Set;
 // CRUD refers Create, Read, Update, Delete
 
 // https://docs.spring.io/spring-data/jpa/docs/1.5.0.RELEASE/reference/html/jpa.repositories.html
-public interface FilmRepository extends CrudRepository<Film, Integer> {
+public interface FilmRepository extends CrudRepository<Film, Long> {
 //    jpql
 
 public Set<Film> getFilmsByTitle(String searchTitle);
